% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add1_stdmod.R
\name{add1.std_selected}
\alias{add1.std_selected}
\title{add1 method for \code{std_selected} class output}
\usage{
\method{add1}{std_selected}(object, ...)
}
\arguments{
\item{object}{The output of the class \code{std_selected}.}

\item{...}{Optional arguments. They will be ignored.}
}
\description{
Intercept the add1 method and raise an error.
}
\details{
\code{\link[=add1]{add1()}} should be not be used after an \code{lm} object is processed by
\code{\link[=std_selected]{std_selected()}} or \code{\link[=std_selected_boot]{std_selected_boot()}}.
}
\examples{
# See examples for std_selected.
}
