% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_condeff.R
\name{coef.cond_effect}
\alias{coef.cond_effect}
\title{Conditional Effect in a 'cond_effect'-Class Object}
\usage{
\method{coef}{cond_effect}(object, ...)
}
\arguments{
\item{object}{The output of \code{\link[=cond_effect]{cond_effect()}} or \code{\link[=cond_effect_boot]{cond_effect_boot()}}.}

\item{...}{Optional arguments. Ignored by the function.}
}
\value{
A numeric vector: The estimates of the conditional
effects in a \code{cond_effect}-class object.
}
\description{
Return the estimates of the conditional
effects
in the output of
\code{\link[=cond_effect]{cond_effect()}} or \code{\link[=cond_effect_boot]{cond_effect_boot()}}.
}
\details{
It just extracts and returns the column of
conditional effects in a \code{cond_effect}-class object.
}
\examples{

# To Do

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
