% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_condeff.R
\name{confint.cond_effect}
\alias{confint.cond_effect}
\title{Confidence Intervals for a 'cond_effect' Class Object}
\usage{
\method{confint}{cond_effect}(object, parm, level = 0.95, type, ...)
}
\arguments{
\item{object}{The output of \code{\link[=cond_effect]{cond_effect()}} or \code{\link[=cond_effect_boot]{cond_effect_boot()}}.}

\item{parm}{Ignored by this function. The confidence intervals for all
available levels will be returned.}

\item{level}{The level of confidence. For the confidence intervals returned
by \code{\link[=lm]{lm()}}, default is .95, i.e., 95\%. For the bootstrap
percentile confidence
intervals, default is the level used in calling
\code{\link[=cond_effect_boot]{cond_effect_boot()}}.}

\item{type}{The type of the confidence intervals. If est to \code{"lm"},
returns the confidence interval given by the \code{\link[=confint]{confint()}} method
of \code{\link[=lm]{lm()}}.
If set to \code{"boot"}, the bootstrap percentile confidence intervals
are returned.
Default is \code{"boot"} if bootstrap estimates are stored in
\code{object}, and \code{"lm"} if bootstrap estimates are not stored.}

\item{...}{Additional arguments. Ignored.}
}
\value{
A matrix of the confidence intervals.
}
\description{
Return the confidence intervals of estimates
conditional effect in the output of
\code{\link[=cond_effect]{cond_effect()}} or \code{\link[=cond_effect_boot]{cond_effect_boot()}}.
}
\details{
If bootstrapping is used to form the confidence interval by
\code{\link[=cond_effect_boot]{cond_effect_boot()}},
users can request the percentile confidence intervals of
the bootstrap estimates. This method does not do the bootstrapping itself.
}
\examples{

# TODO

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
