% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stdmod.R
\name{print.std_selected}
\alias{print.std_selected}
\title{Print method for \code{std_selected} class output}
\usage{
\method{print}{std_selected}(x, ...)
}
\arguments{
\item{x}{The output of the class \code{std_selected}.}

\item{...}{Arguments to be passed to \code{print.lm}.}
}
\value{
Nothing
}
\description{
Print method for \code{std_selected} class output
}
\examples{
# See examples for std_selected.
}
