% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_stdmod.R
\name{print.summary.stdmod}
\alias{print.summary.stdmod}
\title{Print method for \code{summary.stdmod} class output}
\usage{
\method{print}{summary.stdmod}(x, ...)
}
\arguments{
\item{x}{The output of the class \code{summary.stdmod}.}

\item{...}{Arguments to be passed to \code{summary.lm}.}
}
\value{
Nothing
}
\description{
Print method for \code{summary.stdmod} class output
}
\details{
Print method for \code{summary.stdmod} class output
}
\examples{
# See examples for std_selected.
}
