% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleep_emo_con}
\alias{sleep_emo_con}
\title{A real dataset of 500 Hong Kong adults with the following variables
case_id
sleep_duration (in hours)
conscientiousness
emotional_stability}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 500 rows and 6 columns.
}
\usage{
sleep_emo_con
}
\description{
A real dataset of 500 Hong Kong adults with the following variables
case_id
sleep_duration (in hours)
conscientiousness
emotional_stability
}
\keyword{datasets}
