% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_selected_boot.R
\name{std_selected_boot}
\alias{std_selected_boot}
\title{Compute the standardized moderation effect and betas for other predictors given the \code{lm} output, with bootstrapping CIs}
\usage{
std_selected_boot(
  lm_out,
  ...,
  nboot = 100,
  boot_args = NULL,
  full_output = FALSE
)
}
\arguments{
\item{lm_out}{The output from \code{lm}.}

\item{...}{Arguments to be passed to \code{std_selected}.}

\item{nboot}{The number of bootstrap samples. Default is 100.}

\item{boot_args}{A named list of arguments to be passed to \code{boot}. Default
is NULL.}

\item{full_output}{Whether the full output from \code{boot} is return. Default is
FALSE.}
}
\value{
The updated \code{lm} output, with the class \code{stdmod} added. It will be
treated as a usual \code{lm} object by most functions. It has these additional elements:
\itemize{
\item \code{scaled}: The terms scaled.
\item \code{centered}: The terms centered.
\item \code{scaled_by}: The scaling factors. The value is 1 for terms not scaled.
\item \code{centered}: The values used for centering. The value is 0 for terms not centered.
}
}
\description{
Compute the standardized moderation effect and betas for other predictors given the \code{lm} output, with bootstrapping CIs
}
\details{
Compute the standardized moderation effect and betas for other predictors given the \code{lm} output, with bootstrapping CIs
}
\examples{
# "To be prepared"
}
