% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_selected_lavaan.R
\name{std_selected_lavaan}
\alias{std_selected_lavaan}
\title{Standardize coefficients in a lavaan-Model}
\usage{
std_selected_lavaan(
  object,
  to_standardize = ".all.",
  not_to_standardize = NULL,
  skip_categorical_x = TRUE,
  output = c("data.frame", "text"),
  std_se = c("none", "delta", "bootstrap"),
  std_z = TRUE,
  std_pvalue = TRUE,
  std_ci = TRUE,
  level = 0.95,
  progress = TRUE,
  ...,
  delta_method = c("numDeriv", "lavaan")
)
}
\arguments{
\item{object}{The output of
\code{lavaan} model fit functions, such
as \code{\link[lavaan:sem]{lavaan::sem()}} and \code{\link[lavaan:cfa]{lavaan::cfa()}}.}

\item{to_standardize}{A string vector,
which should be the names of the
variables to be standardized.
Default is \code{".all"}, indicating all
variables are to be standardized
(but see \code{skip_categorical}).}

\item{not_to_standardize}{A string
vector, which should be the names
of the variables that should not be
standardized. This argument is useful
when most variables, except for a few,
are to be standardized. This argument
cannot be ued with \code{to_standardize}
at the same time. Default is \code{NULL},
and only \code{to_standardize} is used.}

\item{skip_categorical_x}{Logical.
If \code{TRUE}, the default, all
categorical predictors, defined as
variables with only two possible
values in the data analyzed, will
be skipped in standardization. This
overrides the argument
\code{to_standardize}. That is, a
categorical predictor will not be
standardized even if listed in
\code{to_standardize}, unless uses set
this argument to \code{FALSE}.}

\item{output}{The format of the
output. If set to \code{"data.frame"},
the output will be a data frame like
the one from \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}.
If set to \code{"text"}, it will be printed
in a format similar to the one in
the \code{summary}-method of \code{lavaan}.
(NOTE: \code{"text"} supported but not
fully implemented.)}

\item{std_se}{String. If set to \code{"none"},
the default, standard errors will not
be computed for the standardized
solution. If set to \code{"delta"},
delta method will be used to compute
the standard errors. If est to
\code{"bootstrap"} \emph{and} bootstrap
estimates are available in \code{object}
(e.g., bootstrapping is requested
when fitting the model in \code{lavaan}),
then bootstrap standard errors and
confidence intervals will be
computed. (NOTE: Bootstrap SEs
and CIs are not yet available.)}

\item{std_z}{Logical. If \code{TRUE} and
\code{std_se} is not set to \code{"none"},
standard error will be computed
using the method specified in
\code{std_se}. Default is \code{TRUE}.}

\item{std_pvalue}{Logical. If \code{TRUE},
\code{std_se} is not set to \code{"none"},
and \code{std_z} is \code{TRUE}, \emph{p}-values
will be computed using the method
specified in \code{std_se}. For
bootstrapping, the method proposed by
Asparouhov and Muthén (2021) is used.
Default is \code{TRUE}.}

\item{std_ci}{Logical. If \code{TRUE} and
\code{std_se} is not set to \code{"none"},
confidence intervals will be
computed using the method specified in
\code{std_se}. Default is \code{FALSE.}}

\item{level}{The level of confidence
of the confidence intervals. Default
is .95. It will be used in the
confidence intervals of both
the unstandardized and
standardized solution.}

\item{progress}{Logical. If \code{TRUE},
progress bars will be displayed
for long process.}

\item{...}{Optional arguments to be
passed to the \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}},
which will be use to generate the
output.}

\item{delta_method}{The method used
to compute delta method standard
errors. For internal use and should
not be changed.}
}
\value{
A data frame storing the parameter
estimates, similar in form to the
output of \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}.
}
\description{
Can standardize selected
variables in a \code{lavaan} model without
refitting the models, can handle
product term correctly and skip
categorical predictors in
standardization.
}
\details{
This function lets users
select which variables to be standardized
when computing the standardized
solution. It has the following
features:
\itemize{
\item It automatically skips predictors
which has only two unique values,
assuming that they are dummy variables.
\item It does not standardize product
term, which is incorrect. Instead,
it compute the product term with
its component variables standardized.
\item It can be used to generate bootstrap
confidence intervals for the
standardized solution. Bootstrap
confidence interval is better than
doing standardization \emph{before} fitting
a model because it correctly takes
into account the sampling variance
of the standard deviations. It is
also better than delta method
confidence interval because it takes
into account the usually asymmetric
distribution of parameters after
standardization.
\item For comparison, it can also report
delta method standard errors and
confidence intervals.
}
\subsection{Problems With Common Approaches}{

In most SEM programs, users
have limited control on which
variables to standardize when
requesting the standardized solution.
The solution may be uninterpretable
or misleading in these conditions:
\itemize{
\item Dummy variables are standardized
and can not be interpreted as the
difference between two groups on the
outcome variables.
\item Product terms (interaction terms)
are standardized and they cannot be
interpreted as the changes in the
effects of focal variables when the
moderators change (Cheung, Cheung,
Lau, Hui, & Vong, 2022).
\item Variables with meaningful units can
be more difficult to interpret when
they are standardized (e.g., age).
}

Moreover, the delta method is usually
used, which is suboptimal for
standardization unless the sample
size is large. For example, the
covariance with variables standardized
is a correlation, and its sampling
distribution is skewed unless its
population value is zero. However,
delta method confidence interval
for the correlation is necessarily
symmetric around the point estimate.
}

\subsection{Limitations}{
\itemize{
\item It only supports observed variable
interaction terms, and only support
two-way interactions.
\item It only supports models fitted to
raw data.
\item It does not yet support ordinal
endogenous variables (e.g.,
ordinal indicators).
\item Intercepts not supported.
\item User-defined parameters not
supported. Users should take care of
appropriate standardization themselves
when defining the user parameters.
}
}
}
\examples{
\donttest{
# TO ADD
}

}
\references{
Asparouhov, A., & Muthén, B. (2021). Bootstrap p-value computation.
Retrieved from https://www.statmodel.com/download/FAQ-Bootstrap\%20-\%20Pvalue.pdf

Cheung, S. F., Cheung, S.-H., Lau, E. Y. Y., Hui, C. H., & Vong, W. N.
(2022) Improving an old way to measure moderation effect in standardized
units. \emph{Health Psychology}, \emph{41}(7), 502-505.
\doi{10.1037/hea0001188}
}
\seealso{
\code{\link[=std_selected]{std_selected()}} and \code{\link[=std_selected_boot]{std_selected_boot()}}
for multiple regression conducted by \code{\link[=lm]{lm()}}.
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
