% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmod-package.R
\docType{package}
\name{stdmod-package}
\alias{stdmod-package}
\title{stdmod: Standardized Moderation Effect and Its Confidence Interval}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions for computing a standardized moderation effect in moderated regression and forming its confidence interval by nonparametric bootstrapping as proposed in Cheung, Cheung, Lau, Hui, and Vong (2022) \doi{10.1037/hea0001188}. Also includes simple-to-use functions for computing conditional effects (unstandardized or standardized) and plotting moderation effects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/stdmod/}
  \item Report bugs at \url{https://github.com/sfcheung/stdmod/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})
}

Other contributors:
\itemize{
  \item David Weng Ngai Vong [contributor]
}

}
\keyword{internal}
