% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmod-package.R, R/stdmod.R
\docType{package}
\name{stdmod}
\alias{stdmod}
\title{stdmod: Helper functions for standardized moderation effects}
\usage{
stdmod(
  lm_out,
  x = NULL,
  w = NULL,
  y = NULL,
  x_rescale = TRUE,
  w_rescale = TRUE,
  y_rescale = TRUE
)
}
\arguments{
\item{lm_out}{The output from \code{lm}.}

\item{x}{The independent variable, that is, the variable with its effect
being moderated. If supplied, it's standard deviation will be used
for rescaling. Default is NULL.}

\item{w}{The moderator. If supplied, it's standard deviation will be used
for rescaling. Default is NULL.}

\item{y}{The dependent (outcome) variable. If supplied, it's standard
deviation will be used for rescaling. Default is NULL.}

\item{x_rescale}{If TRUE, will rescale x by its SD. Default is TRUE.}

\item{w_rescale}{If TRUE, will rescale w by its SD. Default is TRUE.}

\item{y_rescale}{If TRUE, will rescale y by its SD. Default is TRUE.}
}
\value{
The standardized moderation effect.
}
\description{
Compute the standardized moderation effect given the \code{lm} output.
}
\details{
Compute the standardized moderation effect given the \code{lm} output.
}
\examples{
# "To be prepared"
}
