% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmod_bootci.R
\name{stdmod_bootci}
\alias{stdmod_bootci}
\title{Compute the bootstrapping CI of the standardized moderation effect given the \code{lm} output}
\usage{
stdmod_bootci(lm_out, ..., nboot = 100, boot_args = NULL, full_output = FALSE)
}
\arguments{
\item{lm_out}{The output from \code{lm}.}

\item{...}{Parameters to be passed to stdmod}

\item{nboot}{The number of bootstrapping samples. Default is 100}

\item{boot_args}{A named list of arguments to be passed to \code{boot}. Default
is NULL.}

\item{full_output}{Whether the full output from \code{boot} is return. Default is
FALSE.}
}
\value{
The standardized moderation effect.
}
\description{
Compute the bootstrapping CI of the standardized moderation effect given the \code{lm} output
}
\details{
Compute the bootstrapping CI of the standardized moderation effect given the \code{lm} output
}
\examples{
# "To be prepared"
}
