% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stdmod.R
\name{summary.std_selected}
\alias{summary.std_selected}
\title{Summary method for \code{std_selected} class output}
\usage{
\method{summary}{std_selected}(object, ...)
}
\arguments{
\item{object}{The output of the class \code{std_selected}.}

\item{...}{Arguments to be passed to \code{summary.lm}.}
}
\value{
An object of class \code{summary.std_selected}, with bootstrapping confidence intervals
added if present.
}
\description{
Summary method for \code{std_selected} class output
}
\details{
Summary method for \code{std_selected} class output
}
\examples{
# See examples for std_selected.
}
