% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stdmod.R
\name{summary.stdmod}
\alias{summary.stdmod}
\title{Summary method for \code{stdmod} class output}
\usage{
\method{summary}{stdmod}(object, ...)
}
\arguments{
\item{object}{The output of the class \code{stdmod}.}

\item{...}{Arguments to be passed to \code{summary.lm}.}
}
\value{
An object of class \code{summary.stdmod}, with bootstrapping confidence intervals
added if present.
}
\description{
Summary method for \code{stdmod} class output
}
\details{
Summary method for \code{stdmod} class output
}
\examples{
# "To be prepared"
}
