% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_3.R
\docType{data}
\name{test_mod1}
\alias{test_mod1}
\title{Sample Dataset: A Path Model With A Moderator}
\format{
A data frame with 300 rows and 6 variables:
\describe{
\item{dv}{Dependent variable, continuous}
\item{iv}{Independent variable, continuous}
\item{med}{Mediator, continuous}
\item{mod}{Moderator, continuous}
\item{cov1}{Covariate, continuous}
\item{cov2}{Covariate, continuous}
}
}
\usage{
test_mod1
}
\description{
For testing. Generated from the following
model.\preformatted{mod <-
"
med ~ iv + mod + iv:mod + cov1
dv ~ med + cov2
"
}
}
\keyword{datasets}
