% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_4.R
\docType{data}
\name{test_mod3_miss}
\alias{test_mod3_miss}
\title{Sample Dataset: A Path Model With A Moderator}
\format{
A data frame with 500 rows and 6 variables:
\describe{
\item{dv}{Dependent variable, continuous}
\item{iv}{Independent variable, continuous}
\item{med}{Mediator, continuous}
\item{mod}{Moderator, continuous}
\item{cov1}{Covariate, continuous}
\item{cov2}{Covariate, continuous}
}
}
\usage{
test_mod3_miss
}
\description{
For testing the handling of warnings
in \code{\link[=stdmod_lavaan]{stdmod_lavaan()}}. Generated from the following
model. \code{dv} has about 88\% missing. A warning on missing
data will be raised in some bootstrap samples.\preformatted{mod <-
"
med ~ iv + mod + iv:mod + cov1
dv ~ med + cov2
"
}
}
\keyword{datasets}
