% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{test_x_1_w_1_v_1_cat1_xw_cov_n_500}
\alias{test_x_1_w_1_v_1_cat1_xw_cov_n_500}
\title{One IV, one moderation, two covariates, v1 is continuous
and cat2 is categorical. IV and moderator have a nonlinear
relation.}
\format{
An object of class \code{data.frame} with 500 rows and 5 columns.
}
\usage{
test_x_1_w_1_v_1_cat1_xw_cov_n_500
}
\description{
Sample size is 500.
}
\keyword{datasets}
