% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_x_1_w_1_v_2_n_100}
\alias{test_x_1_w_1_v_2_n_100}
\title{Sample dataset for testing}
\format{
An object of class \code{data.frame} with 500 rows and 5 columns.
}
\usage{
test_x_1_w_1_v_2_n_100
}
\description{
One IV, one moderation, two covariates, n 100
}
\keyword{datasets}
