% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stdmod.R
\name{update.std_selected}
\alias{update.std_selected}
\title{update method for \code{std_selected} class output}
\usage{
\method{update}{std_selected}(object, ...)
}
\arguments{
\item{object}{The output of the class \code{std_selected}.}

\item{...}{Arguments to be passed to \code{\link[=update]{update()}}.}
}
\value{
An object of class `std_selected``.
}
\description{
Like update for \code{\link[=lm]{lm()}} but will also update scaling and centering.
}
\details{
Only accepts update for \code{\link[=lm]{lm()}}.
}
\examples{
# See examples for std_selected.
}
