% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condeff.R
\name{cond_effect}
\alias{cond_effect}
\title{Conditional Effects}
\usage{
cond_effect(
  output,
  x = NULL,
  w = NULL,
  w_method = c("sd", "percentile"),
  w_percentiles = c(0.16, 0.84),
  w_sd_to_percentiles,
  w_from_mean_in_sd = 1
)
}
\arguments{
\item{x}{The independent variable, that is, the variable with its effect
being moderated. If supplied, its standard deviation will be
used
for rescaling. Default is \code{NULL}.}

\item{w}{The moderator. If supplied, its standard deviation will be
used
for rescaling. Default is \code{NULL}.}

\item{w_method}{How to define "high" and "low" for the moderator levels.
Default is in terms of the
standard deviation of the moderator, \verb{"sd". If equal to }"percentile"\verb{, then percentiles of the moderator in the dataset are used. Ignored if }w` is categorical.}

\item{w_percentiles}{If \code{w_method} is \code{"percentile"}, then this
argument
specifies the two percentiles to be used,
divided by 100.
It must be a
vector of two numbers. The default is
\code{c(.16, .84)},
the 16th and 84th percentiles,
which corresponds approximately
to one SD below and above mean for a
normal distribution, respectively.
Ignored if \code{w} is categorical.}

\item{w_sd_to_percentiles}{If \code{w_method} is \code{"percentile"} and
this argument is
set to a number, this number will be
used to
to determine the percentiles to be used.
The
lower percentile is the percentile in a
normal
distribution
that is \code{w_sd_to_percentiles} SD below
the mean.
The upper percentile is the percentile in
a normal
distribution that is \code{w_sd_to_percentiles}
SD
above the mean. Therefore, if
\code{w_sd_to_percentiles} is set to 1, then the
lower
and upper percentiles are 16th and 84th,
respectively.}

\item{w_from_mean_in_sd}{How many SD from mean is used to define
"low" and
"high" for the moderator. Default is 1.
Ignored if \code{w} is categorical.}

\item{lm_out}{The output from \code{\link[=lm]{lm()}}.}
}
\value{
The data frame of conditional effects.
}
\description{
Compute the conditional effects in a moderated regression model.
}
\examples{
# TODO

}
\references{
Cheung, S. F., Cheung, S.-H., Lau, E. Y. Y., Hui, C. H., & Vong, W. N.
(2022) Improving an old way to measure moderation effect in standardized
units. Advance online publication. \emph{Health Psychology}.
\doi{10.1037/hea0001188}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
