% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condeff.R
\name{cond_effect}
\alias{cond_effect}
\title{Conditional Effects}
\usage{
cond_effect(
  output,
  x = NULL,
  w = NULL,
  w_method = c("sd", "percentile"),
  w_percentiles = c(0.16, 0.5, 0.84),
  w_sd_to_percentiles = NA,
  w_from_mean_in_sd = 1
)
}
\arguments{
\item{output}{The output from \code{\link[stats:lm]{stats::lm()}}. It can also accept the output
from
\code{\link[=std_selected]{std_selected()}} or \code{\link[=std_selected_boot]{std_selected_boot()}}.}

\item{x}{The independent variable, that is, the variable with its effect
being moderated. It must be a numeric variable.}

\item{w}{The moderator. Both numeric variables and categorical variables
(character or factor) are supported.}

\item{w_method}{How to define "low", "medium", and "high" for the moderator
levels.
Default is in terms of mean and
standard deviation (SD) of the moderator, \code{"sd"}:
"low", "medium", and "high" are one SD below mean,
mean, and one SD above mean, respectively.
If equal to
\code{"percentile"}, then percentiles of the moderator in
the
dataset are used: "low", "medium", and "high" are
16th, 50th (median), and 84th percentiles, respectively.
Ignored if \code{w} is categorical.}

\item{w_percentiles}{If \code{w_method} is \code{"percentile"}, then this
argument
specifies the three percentiles to be used,
divided by 100.
It must be a
vector of two numbers. The default is
\code{c(.16, .50, .84)},
the 16th, 50th, and 84th percentiles,
which corresponds approximately
to one SD below and above mean in a
normal distribution, respectively.
Ignored if \code{w} is categorical.}

\item{w_sd_to_percentiles}{If \code{w_method} is \code{"percentile"} and
this argument is
set to a number, this number will be
used to
to determine the percentiles to be used.
The
lower percentile is the percentile in a
normal
distribution
that is \code{w_sd_to_percentiles} SD below
the mean.
The upper percentile is the percentile in
a normal
distribution that is \code{w_sd_to_percentiles}
SD
above the mean. Therefore, if
\code{w_sd_to_percentiles} is set to 1, then the
lower
and upper percentiles are 16th and 84th,
respectively. Default is \code{NA}.}

\item{w_from_mean_in_sd}{How many SD from mean is used to define
"low" and
"high" for the moderator. Default is 1.
Ignored if \code{w} is categorical.}
}
\value{
A data-frame-like object of the conditional effects. The class is
\code{cond_effect} and the print method will print additional information of
the conditional effects.
}
\description{
Compute the conditional effects in a moderated regression model.
}
\examples{

# Load a sample data set
# It has one predictor (iv), one moderator (mod), on covariate (v1),
# one categorical covariate (cat1) with three groups, and one dv (dv).
dat <- test_x_1_w_1_v_1_cat1_n_500

# Do a moderated regression by lm
lm_raw <- lm(dv ~ iv*mod + v1 + cat1, dat)
summary(lm_raw)
cond_effect(lm_raw, x = iv, w = mod)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
