% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_stdmod.R
\name{confint.std_selected}
\alias{confint.std_selected}
\title{confint method for \code{std_selected} class output}
\usage{
\method{confint}{std_selected}(object, parm, level = 0.95, type = "lm", ...)
}
\arguments{
\item{object}{The output of the class \code{std_selected}.}

\item{parm}{The parameters for which confidence intervals should be returned.
If missing, the confidence intervals of all parameters will be
returned.}

\item{level}{The level of confidence. For the confidence intervals returned
by \link{lm}, default is .95. For the bootstrapping confidence
intervals, default is the level used in calling \link{std_selected_boot}.}

\item{type}{The type of the confidence intervals. Default is "lm",
returned by the \link{confint} method of \link{lm}. If set to "boot",
the bootstrap percentile confidence intervals are
returned.}

\item{...}{Arguments to be passed to \code{summary.lm}.}
}
\value{
A matrix of the confidence intervals.
}
\description{
Return the confidence intervals of estimates
in a \code{std_selected} class output
}
\details{
If bootstrapping is used to form the confidence interval,
users can request the percentile confidence intervals of
the bootstrap estimates.
}
\examples{
# See examples for std_selected.
}
