% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmod.R
\name{plotmod}
\alias{plotmod}
\title{Moderation Effect Plot}
\usage{
plotmod(
  output,
  x,
  w,
  x_label,
  w_label,
  y_label,
  title,
  digits = 3,
  x_from_mean_in_sd = 1,
  w_from_mean_in_sd = 1,
  w_method = c("sd", "percentile"),
  w_percentiles = c(0.16, 0.84),
  x_method = c("sd", "percentile"),
  x_percentiles = c(0.16, 0.84),
  w_sd_to_percentiles,
  x_sd_to_percentiles,
  note_standardized = TRUE,
  no_title = FALSE
)
}
\arguments{
\item{output}{An object of the class \code{lm}, such as the output
of \code{\link[stats:lm]{stats::lm()}}, \code{\link[=std_selected]{std_selected()}}, or
\code{\link[=std_selected_boot]{std_selected_boot()}}.}

\item{x}{The name of the focal variable (x-axis) in the output. It
can be the name of the variable, with or without quotes.
Currently only numeric variables are supported.}

\item{w}{The name of the moderator in the output. It
can be the name of the variable, with or without quotes.}

\item{x_label}{The label for the X-axis. Default is the value of \code{x}.}

\item{w_label}{The label for the legend for the lines.
Default is the value of\code{w}.}

\item{y_label}{The label for the Y-axis. Default is the
name of the response variable in the model.}

\item{title}{The title of the graph. If not supplied, will be
generated from the variable
names or labels (in \code{x_label}, \code{y_label},
and \code{w_label}). If \code{""}, no title will be printed.
This can be used when the plot is for manuscript
submission and figures are required to have no
titles.}

\item{digits}{Number of decimal digits to print. Default is 3.}

\item{x_from_mean_in_sd}{How many SD from mean is used to define
"low" and
"high" for the focal variable.
Default is 1.}

\item{w_from_mean_in_sd}{How many SD from mean is used to define
"low" and
"high" for the moderator. Default is 1.
Ignored if \code{w} is categorical.}

\item{w_method}{How to define "high" and "low" for the moderator levels.
Default is in terms of the
standard deviation of the moderator, \verb{"sd". If equal to }"percentile"\verb{, then percentiles of the moderator in the dataset are used. Ignored if }w` is categorical.}

\item{w_percentiles}{If \code{w_method} is \code{"percentile"}, then this
argument
specifies the two percentiles to be used,
divided by 100.
It must be a
vector of two numbers. The default is
\code{c(.16, .84)},
the 16th and 84th percentiles,
which corresponds approximately
to one SD below and above mean for a
normal distribution, respectively.
Ignored if \code{w} is categorical.}

\item{x_method}{How to define "high" and "low" for the focal
variable levels.
Default is in terms of the
standard deviation of the focal variable, \code{"sd"}.
If equal to
\code{"percentile"}, then the percentiles of the
focal variable in the
dataset is used.}

\item{x_percentiles}{If \code{x_method} is \code{"percentile"}, then this
argument
specifies the two percentiles to be used,
divided by 100.
It must be a
vector of two numbers. The default is
\code{c(.16, .84)},
the 16th and 84th percentiles,
which corresponds approximately
to one SD below and above mean for a
normal distribution, respectively.}

\item{w_sd_to_percentiles}{If \code{w_method} is \code{"percentile"} and
this argument is
set to a number, this number will be
used to
to determine the percentiles to be used.
The
lower percentile is the percentile in a
normal
distribution
that is \code{w_sd_to_percentiles} SD below
the mean.
The upper percentile is the percentile in
a normal
distribution that is \code{w_sd_to_percentiles}
SD
above the mean. Therefore, if
\code{w_sd_to_percentiles} is set to 1, then the
lower
and upper percentiles are 16th and 84th,
respectively.}

\item{x_sd_to_percentiles}{If \code{x_method} is \code{"percentile"} and this
argument is
set to a number, this number will be used
to
to determine the percentiles to be used. The
lower percentile is the percentile in a
normal
distribution
that is \code{x_sd_to_percentiles} SD below the mean.
The upper percentile is the percentile in a normal
distribution that is \code{x_sd_to_percentiles} SD
above the mean. Therefore, if
\code{x_sd_to_percentiles} is set to 1, then the lower
and upper percentiles are 16th and 84th,
respectively.}

\item{note_standardized}{If \code{TRUE}, will check whether a variable has SD
nearly equal to one. If yes, will report this in the
plot.
Default is \code{TRUE}.}

\item{no_title}{If \code{TRUE}, title will be suppressed. Default is \code{FALSE}.}
}
\value{
A \link{ggplot2} graph. Plotted if not assigned to a name. It can
be further modified like a usual \link{ggplot2} graph.
}
\description{
Plot the moderation effect in a regression model
}
\details{
This function generate a basic \link{ggplot2} graph
typically found psychology manuscripts. It tries to
check whether one or more variables are standardized, and
report this in the plot if required.

This function only has features for typical plots of moderation effects.
It is not intended to be a flexible tool for a fine control on the plots.
}
\examples{

# Do a moderated regression by lm
lm_out <- lm(sleep_duration ~ age + gender + emotional_stability*conscientiousness, sleep_emo_con)
summary(lm_out)

# Standardize all variables except for categorical variables
lm_std <- std_selected(lm_out,
                       to_scale = ~ .,
                       to_center = ~ .)
summary(lm_std)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
