% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cond_eff.R
\name{print.cond_effect}
\alias{print.cond_effect}
\title{Print the Summary of a 'cond_effect' Class Object}
\usage{
\method{print}{cond_effect}(x, nd = 3, nd_stat = 3, nd_p = 3, ...)
}
\arguments{
\item{x}{The output of \code{\link[=cond_effect]{cond_effect()}}.}

\item{nd}{The number of digits for the variables.}

\item{nd_stat}{The number of digits for test statistics (e.g., t).}

\item{nd_p}{The number of digits for p-values.}

\item{...}{Arguments to be passed to \code{\link[=cond_effect]{cond_effect()}}.}
}
\value{
Nothing
}
\description{
Print the output of \code{\link[=cond_effect]{cond_effect()}}.
}
\examples{

# Load a sample data set
# It has one predictor (iv), one moderator (mod), on covariate (v1),
# one categorical covariate (cat1) with three groups, and one dv (dv).
dat <- test_x_1_w_1_v_1_cat1_n_500

# Do a moderated regression by lm
lm_raw <- lm(dv ~ iv*mod + v1 + cat1, dat)
summary(lm_raw)
cond_effect(lm_raw, x = iv, w = mod)

lm_std <- std_selected(lm_raw, to_scale = ~ iv + mod, to_center = ~ iv + mod)
cond_effect(lm_std, x = iv, w = mod)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
