% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleep_emo_con}
\alias{sleep_emo_con}
\title{Sample Dataset: Predicting Sleep Duration}
\format{
A data frame with 500 rows and six variables:
\describe{
\item{case_id}{Case ID, integer}
\item{sleep_duration}{Sleep duration in hours}
\item{conscientiousness}{Conscientiousness score, continuous}
\item{emotional_stability}{Emotional stability score, continuous}
\item{age}{Age in years}
\item{gender}{Gender, string, "female" or "male"}
}
}
\usage{
sleep_emo_con
}
\description{
A random subset from a real dataset. For
illustration.
}
\keyword{datasets}
