% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_selected.R
\name{std_selected}
\alias{std_selected}
\title{Compute the standardized moderation effect and betas for other predictors given the \code{lm} output}
\usage{
std_selected(lm_out, to_scale = NULL, to_center = NULL)
}
\arguments{
\item{lm_out}{The output from \code{lm}.}

\item{to_scale}{Specify the terms to be rescaled by standard deviation,
using a formula as in \code{lm}. For example, if the terms to be scale
is x1 and x3, use \code{~ x1 + x3}. No need to specify the interaction term.
Specify only the original variables. If NULL, then all terms
will be rescaled by standard deviation. Variables that are not numeric will
will be ignored. Default is NULL.}

\item{to_center}{Specify the terms to be mean-centered, using a formula
as in \code{lm}. For example, if the terms to be scale
is x1 and x3, use \code{~ x1 + x3}. No need to specify the interaction term.
Specify only the original variables. If NULL, then all terms
will be centered. Default is NULL.}
}
\value{
The updated \code{lm} output, with the class \code{stdmod} added. It will be
treated as a usual \code{lm} object by most functions. It has these additional elements:
\itemize{
\item \code{scaled}: The terms scaled.
\item \code{centered}: The terms centered.
\item \code{scaled_by}: The scaling factors. The value is 1 for terms not scaled.
\item \code{centered}: The values used for centering. The value is 0 for terms not centered.
}
}
\description{
Compute the standardized moderation effect and betas for other predictors given the \code{lm} output
}
\details{
Compute the standardized moderation effect and betas for other predictors given the \code{lm} output
}
\examples{
# "To be prepared"
}
