% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_selected.R, R/std_selected_boot.R
\name{std_selected}
\alias{std_selected}
\alias{std_selected_boot}
\title{Compute the standardized moderation effect and betas for other predictors given the \code{lm} output}
\usage{
std_selected(lm_out, to_scale = NULL, to_center = NULL)

std_selected_boot(
  lm_out,
  ...,
  nboot = 100,
  boot_args = NULL,
  full_output = FALSE
)
}
\arguments{
\item{lm_out}{The output from \code{lm}.}

\item{to_scale}{Specify the terms to be rescaled by standard deviation,
using a formula as in \code{lm}. For example, if the terms to be scale
is x1 and x3, use \code{~ x1 + x3}. No need to specify the interaction term.
Specify only the original variables. If NULL, then no terms
will be rescaled by standard deviation. Variables that are not numeric will
will be ignored. Default is NULL.}

\item{to_center}{Specify the terms to be mean-centered, using a formula
as in \code{lm}. For example, if the terms to be scale
is x1 and x3, use \code{~ x1 + x3}. No need to specify the interaction term.
Specify only the original variables. If NULL, then no terms
will be centered. Default is NULL.}

\item{...}{Arguments to be passed to \code{std_selected}.}

\item{nboot}{The number of bootstrap samples. Default is 100.}

\item{boot_args}{A named list of arguments to be passed to \code{\link[boot:boot]{boot::boot()}}. Default
is \code{NULL}.}

\item{full_output}{Whether the full output from \code{\link[boot:boot]{boot::boot()}} is return. Default is
\code{FALSE}.}
}
\value{
The updated \code{lm} output, with the class \code{stdmod} added. It will be
treated as a usual \code{lm} object by most functions. It has these additional elements:
\itemize{
\item \code{scaled}: The terms scaled.
\item \code{centered}: The terms centered.
\item \code{scaled_by}: The scaling factors. The value is 1 for terms not scaled.
\item \code{centered}: The values used for centering. The value is 0 for terms not centered.
}
}
\description{
Compute the standardized moderation effect and betas for other predictors given the \code{lm} output
}
\details{
\code{\link[=std_selected]{std_selected()}} compute the standardized moderation effect and betas for other predictors given
the \code{lm} output

This function is a wrapper of \code{\link[=std_selected]{std_selected()}}. It calls \code{\link[=std_selected]{std_selected()}} once
for each bootstrap sample, and then compute the nonparametric bootstrapping
confidence intervals.

This function intentionally does not have an argument for setting the seed for
random number. Users are recommended to set the seed, e.g., using \code{\link[=set.seed]{set.seed()}}
before calling it, to ensure reproducibility.
}
\section{Functions}{
\itemize{
\item \code{std_selected}: The base function to center or scale selected variables

\item \code{std_selected_boot}: A wrapper of \code{\link[=std_selected]{std_selected()}} that forms
nonparametrc bootstrapping intervals.
}}

\examples{
# "To be prepared"
# "To be prepared"
}
