% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmod-package.R, R/stdmod.R
\docType{package}
\name{stdmod}
\alias{stdmod}
\title{stdmod: Helper functions for standardized moderation effects}
\usage{
stdmod(lm_out, x = NULL, w = NULL, y = NULL)
}
\arguments{
\item{lm_out}{The output from \code{lm}.}

\item{x}{The independent variable, that is, the variable with its effect
being moderated. If supplied, it's standard deviation will be used
for rescaling. Default is NULL.}

\item{w}{The moderator. If supplied, it's standard deviation will be used
for rescaling. Default is NULL.}

\item{y}{The dependent (outcome) variable. If supplied, it's standard
deviation will be used for rescaling. Default is NULL.}
}
\value{
The standardized moderation effect.
}
\description{
Compute the standardized moderation effect given the \code{lm} output.
}
\details{
Compute the standardized moderation effect given the \code{lm} output.
}
\examples{
# "To be prepared"
}
