% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmod.R, R/stdmod_bootci.R, R/stdmod-package.R
\docType{package}
\name{stdmod}
\alias{stdmod}
\alias{stdmod_bootci}
\title{Compute the standardized moderation effect given the \code{lm} output}
\usage{
stdmod(
  lm_out,
  x = NULL,
  w = NULL,
  y = NULL,
  x_rescale = TRUE,
  w_rescale = TRUE,
  y_rescale = TRUE
)

stdmod_bootci(lm_out, ..., nboot = 100, boot_args = NULL, full_output = FALSE)
}
\arguments{
\item{lm_out}{The output from \code{lm}.}

\item{x}{The independent variable, that is, the variable with its effect
being moderated. If supplied, it's standard deviation will be used
for rescaling. Default is NULL.}

\item{w}{The moderator. If supplied, it's standard deviation will be used
for rescaling. Default is NULL.}

\item{y}{The dependent (outcome) variable. If supplied, it's standard
deviation will be used for rescaling. Default is NULL.}

\item{x_rescale}{If TRUE, will rescale x by its SD. Default is TRUE.}

\item{w_rescale}{If TRUE, will rescale w by its SD. Default is TRUE.}

\item{y_rescale}{If TRUE, will rescale y by its SD. Default is TRUE.}

\item{...}{Parameters to be passed to stdmod}

\item{nboot}{The number of bootstrapping samples. Default is 100}

\item{boot_args}{A named list of arguments to be passed to \code{boot}. Default
is NULL.}

\item{full_output}{Whether the full output from \code{boot} is return. Default is
FALSE.}
}
\value{
The standardized moderation effect.
}
\description{
Compute the standardized moderation effect given the \code{lm} output.
}
\details{
Compute the standardized moderation effect given the \code{lm} output.

\code{\link[=stdmod_bootci]{stdmod_bootci()}} is a wrapper of \code{\link[=stdmod]{stdmod()}}. It computes the bootstrapping
confidence interval of the standardized moderation effect given
the \code{lm} output.
}
\section{Functions}{
\itemize{
\item \code{stdmod}: The base function compute standardized moderation effect

\item \code{stdmod_bootci}: A wrapper of \code{\link[=std_mod]{std_mod()}} that compute the nonparametric
bootstrapping confidence interval of the standardized
moderation effect
}}

\examples{
# "To be prepared"
# "To be prepared"
}
