% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmod_lavaan.R
\name{stdmod_lavaan}
\alias{stdmod_lavaan}
\title{Standardized moderation effect and bootstrapping CI in lavaan}
\usage{
stdmod_lavaan(fit, x, y, w, x_w, boot_ci = FALSE, R = 100, conf = 0.95, ...)
}
\arguments{
\item{fit}{The SEM output by \code{\link[lavaan:lavaan]{lavaan::lavaan()}} or its siblings.}

\item{x}{The name of the independent variable in the model.}

\item{y}{The name of the outcome variable in the model.}

\item{w}{The name of the moderator in the model.}

\item{x_w}{The name of the product term (x * w) in the model.}

\item{boot_ci}{Boolean. Whether nonparametric bootstrapping will be
conducted. Default is \code{FALSE}.}

\item{R}{The number of nonparametric bootstrapping samples. Default is 100.
Set this to at least 2000 in actual use.}

\item{conf}{The level of confidence. Default is .95.}

\item{...}{Optional arguments to be passed to \code{\link[boot:boot]{boot::boot()}}.}
}
\value{
A list with these elements:
\itemize{
\item \code{stdmod}: The standardized moderation effect.
\item \code{ci}: The nonparametric bootstrapping confidence interval. \code{NA} if
confidence interval not requested.
\item \code{boot_out}: The raw output from \code{\link[boot:boot]{boot::boot()}}. \code{NA} if
confidence interval not requested.
}
}
\description{
Compute the standardized moderation effect and bootstrapping CI given a lavaan output
}
\details{
\code{\link[=stdmod_lavaan]{stdmod_lavaan()}} accepts a \link[lavaan:lavaan-class]{lavaan::lavaan} object, the SEM output generated
by \code{\link[lavaan:lavaan]{lavaan::lavaan()}} and its sibling (e.g, \code{\link[lavaan:sem]{lavaan::sem()}}) and compute
the standardized moderation effect.

The standard deviations of the independent variable, moderator, and outcome
are computed from the implied covariance matrix returned by \code{\link[lavaan:lavInspect]{lavaan::lavInspect()}}.

If nonparametric bootstrapping confidence is requested with \code{R} bootstrapping
samples, the model will be fitted \code{R} times to these samples, and the standardized
moderation effect will be computed in each sample. This enasures that all
componenets used in the computation, including the standard deviations, are
also computed from the bootstrapping samples.

Note that the compuation can be slow because \code{\link[lavaan:lavaan]{lavaan::lavaan()}} will be called
\code{R} times.
}
\examples{
# "To be prepared"
}
