% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_x_1_w_1_v_1_cat1_n_500}
\alias{test_x_1_w_1_v_1_cat1_n_500}
\title{One IV, one moderation, two covariates, v1 is contrinuous and cat2 is categorical.
Sample size is 500.}
\format{
An object of class \code{data.frame} with 500 rows and 5 columns.
}
\usage{
test_x_1_w_1_v_1_cat1_n_500
}
\description{
One IV, one moderation, two covariates, v1 is contrinuous and cat2 is categorical.
Sample size is 500.
}
\keyword{datasets}
