% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_x_1_w_1_v_1_cat1_n_500}
\alias{test_x_1_w_1_v_1_cat1_n_500}
\title{Sample Dataset: One IV, One Moderator, Two Covariates}
\format{
A data frame with 500 rows and five variables:
\describe{
\item{dv}{Dependent variable, continuous}
\item{iv}{Independent variable, continuous}
\item{mod}{Moderator variable, continuous}
\item{v1}{Covariate, continuous}
\item{cat1}{Covariate, categorical}
}
}
\usage{
test_x_1_w_1_v_1_cat1_n_500
}
\description{
The IV (iv1) continuous; The moderator (cat2) is
categorical. For testing.
}
\keyword{datasets}
