% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stdmod.R
\name{update.std_selected}
\alias{update.std_selected}
\title{update method for \code{std_selected} class output}
\usage{
\method{update}{std_selected}(object, ...)
}
\arguments{
\item{object}{The output of the class \code{std_selected}.}

\item{...}{Optional arguments. They will be ignored.}
}
\description{
Intercept the update method and raise an error.
}
\details{
\code{\link[=update]{update()}} should be done in the call to \code{\link[=lm]{lm()}} before calling
\code{\link[=std_selected]{std_selected()}} or \code{\link[=std_selected_boot]{std_selected_boot()}}.
}
\examples{
# See examples for std_selected.
}
