% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stdmod.R
\name{update.std_selected}
\alias{update.std_selected}
\title{Update method for \code{std_selected} class output}
\usage{
\method{update}{std_selected}(object, formula., ..., evaluate = TRUE)
}
\arguments{
\item{object}{The output of the class \code{std_selected}.}

\item{formula.}{Changes to the formula.}

\item{...}{Optional arguments to be changed.}

\item{evaluate}{Whether the call will be evaluated.}
}
\description{
Update method and raise an error.
}
\details{
This should be used only to update the call to \code{\link[=lm]{lm()}},
not to the call to \code{\link[=std_selected]{std_selected()}} or \code{\link[=std_selected_boot]{std_selected_boot()}}.
}
\examples{
# See examples for std_selected.
}
