% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov_stdmod.R
\name{vcov.std_selected}
\alias{vcov.std_selected}
\title{vcov method for \code{std_selected} class output}
\usage{
\method{vcov}{std_selected}(object, type = "lm", ...)
}
\arguments{
\item{object}{The output of the class \code{std_selected}.}

\item{type}{The type of variance-covariance matrix. Default is "lm",
returned by the \link{vcov} method of \link{lm}. If set to "boot",
the variance-covariance matrix of the bootstrap estimates
is returned.}

\item{...}{Arguments to be passed to \code{summary.lm}.}
}
\value{
A matrix of the variances and covariances.
}
\description{
Compute the variance-covariance matrix
of estimates in a \code{std_selected} class output
}
\details{
If bootstrapping is used to form the confidence interval,
users can request the variance-covariance matrix of the bootstrap estimates.
}
\examples{
# See examples for std_selected.
}
